backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.12
openai==1.57.0
langgraph==0.2.58
langchain==0.3.11
langchain-openai==0.2.12
langchain-community==0.3.11
langchain-text-splitters==0.3.3
faiss-cpu==1.9.0.post1
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.28.1
numpy==2.2.0
sse-starlette==2.1.3
twilio==9.4.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
